refactor: added scss and rename CategoryHeading
