+++ frontend/src/components/Logs/CategoryHeading/CategoryHeading.styles.scss
@@ -1,0 +1,5 @@
+@import '@signozhq/ui-core/packages/token/build/tokens/variables.scss';
+
+.category-heading-text {
+    font-size: $size-font-normal !important;
+}
