--- frontend/src/components/Logs/CategoryHeading/styles.ts
@@ -1,6 +1,0 @@
-import { Typography } from 'antd';
-import styled from 'styled-components';
-
-export const CategoryHeadingText = styled(Typography.Text)`
-	font-size: 0.8rem;
-`;
